+++ src/SimulationTypes/Simulation_StatisticalTrend.cpp
@@ -449,7 +449,7 @@ void Simulation_StatisticalTrend::run()
     extern bool parallelism_enabled;
     #pragma omp parallel for ordered schedule(dynamic) if(parallelism_enabled)
 
-    for (unsigned i = 0; i < simulations.size(); i++)
+    for (size_t i = 0; i < simulations.size(); i++)
         {
         simulations[i]->run();
